Split read and write to own modules
