+++ src/lib.rs
@@ -1,6 +1,7 @@
-use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
 use linked_hash_map::LinkedHashMap;
-use std::io::{Cursor, Read, Write};
+
+pub mod decode;
+pub mod encode;
 
 #[derive(Debug)]
 pub enum TagError<'a> {
@@ -210,236 +211,6 @@ impl CompoundTag {
     }
 }
 
-pub fn read_compound_tag<'a, R: Read>(reader: &mut R) -> Result<CompoundTag, TagError<'a>> {
-    let tag_id = reader.read_u8().unwrap();
-    let (_, tag) = read_tag(tag_id, reader, true)?;
-
-    match tag {
-        Tag::Compound(value) => Ok(value),
-        _ => Err(TagError::ExpectedCompoundTag {
-            unexpected_tag_id: tag.id(),
-        }),
-    }
-}
-
-fn read_tag<'a, R: Read>(
-    id: u8,
-    reader: &mut R,
-    read_name: bool,
-) -> Result<(String, Tag), TagError<'a>> {
-    if id == 0 {
-        return Ok(("".to_owned(), Tag::End));
-    }
-
-    let name = if read_name {
-        read_string(reader)
-    } else {
-        String::from("")
-    };
-
-    match id {
-        1 => {
-            let value = reader.read_i8().unwrap();
-            let tag = Tag::Byte(value);
-
-            return Ok((name, tag));
-        }
-        2 => {
-            let value = reader.read_i16::<BigEndian>().unwrap();
-            let tag = Tag::Short(value);
-
-            return Ok((name, tag));
-        }
-        3 => {
-            let value = reader.read_i32::<BigEndian>().unwrap();
-            let tag = Tag::Int(value);
-
-            return Ok((name, tag));
-        }
-        4 => {
-            let value = reader.read_i64::<BigEndian>().unwrap();
-            let tag = Tag::Long(value);
-
-            return Ok((name, tag));
-        }
-        5 => {
-            let value = reader.read_f32::<BigEndian>().unwrap();
-            let tag = Tag::Float(value);
-
-            return Ok((name, tag));
-        }
-        6 => {
-            let value = reader.read_f64::<BigEndian>().unwrap();
-            let tag = Tag::Double(value);
-
-            return Ok((name, tag));
-        }
-        7 => {
-            let length = reader.read_u32::<BigEndian>().unwrap();
-            let mut value = Vec::new();
-
-            for _ in 0..length {
-                value.push(reader.read_i8().unwrap());
-            }
-
-            let tag = Tag::ByteArray(value);
-
-            return Ok((name, tag));
-        }
-        8 => {
-            let value = read_string(reader);
-            let tag = Tag::String(value);
-
-            return Ok((name, tag));
-        }
-        9 => {
-            let list_tags_id = reader.read_u8().unwrap();
-            let length = reader.read_u32::<BigEndian>().unwrap();
-            let mut value = Vec::new();
-
-            for _ in 0..length {
-                let (_, tag) = read_tag(list_tags_id, reader, false)?;
-                value.push(tag);
-            }
-
-            let tag = Tag::List(value);
-
-            return Ok((name, tag));
-        }
-        10 => {
-            let mut tags = LinkedHashMap::new();
-
-            loop {
-                let tag_id = reader.read_u8().unwrap();
-                let (name, tag) = read_tag(tag_id, reader, true)?;
-
-                match tag {
-                    Tag::End => break,
-                    _ => {
-                        tags.insert(name, tag);
-                    }
-                }
-            }
-
-            let compound_tag = CompoundTag { tags };
-            let tag = Tag::Compound(compound_tag);
-
-            return Ok((name, tag));
-        }
-        11 => {
-            let length = reader.read_u32::<BigEndian>().unwrap();
-            let mut value = Vec::new();
-
-            for _ in 0..length {
-                value.push(reader.read_i32::<BigEndian>().unwrap());
-            }
-
-            let tag = Tag::IntArray(value);
-
-            return Ok((name, tag));
-        }
-        12 => {
-            let length = reader.read_u32::<BigEndian>().unwrap();
-            let mut value = Vec::new();
-
-            for _ in 0..length {
-                value.push(reader.read_i64::<BigEndian>().unwrap());
-            }
-
-            let tag = Tag::LongArray(value);
-
-            return Ok((name, tag));
-        }
-        type_id => return Err(TagError::UnknownType { type_id }),
-    }
-}
-
-fn read_string<R: Read>(reader: &mut R) -> String {
-    let length = reader.read_u16::<BigEndian>().unwrap();
-    let mut buf = vec![0; length as usize];
-    reader.read_exact(&mut buf).unwrap();
-
-    String::from_utf8_lossy(&buf).into_owned()
-}
-
-pub fn write_compound_tag<'a, W: Write>(
-    writer: &mut W,
-    compound_tag: CompoundTag,
-) -> Result<(), TagError<'a>> {
-    write_tag(writer, String::from(""), Tag::Compound(compound_tag), true)
-}
-
-fn write_tag<'a, W: Write>(
-    writer: &mut W,
-    name: String,
-    tag: Tag,
-    write_header: bool,
-) -> Result<(), TagError<'a>> {
-    if write_header {
-        writer.write_u8(tag.id()).unwrap();
-        write_string(writer, name);
-    }
-
-    match tag {
-        Tag::Byte(value) => writer.write_i8(value).unwrap(),
-        Tag::Short(value) => writer.write_i16::<BigEndian>(value).unwrap(),
-        Tag::Int(value) => writer.write_i32::<BigEndian>(value).unwrap(),
-        Tag::Long(value) => writer.write_i64::<BigEndian>(value).unwrap(),
-        Tag::Float(value) => writer.write_f32::<BigEndian>(value).unwrap(),
-        Tag::Double(value) => writer.write_f64::<BigEndian>(value).unwrap(),
-        Tag::ByteArray(value) => {
-            writer.write_u32::<BigEndian>(value.len() as u32).unwrap();
-
-            for v in value {
-                writer.write_i8(v).unwrap();
-            }
-        }
-        Tag::String(value) => write_string(writer, value),
-        Tag::List(value) => {
-            if value.len() > 0 {
-                writer.write_u8(value[0].id()).unwrap()
-            } else {
-                writer.write_u8(Tag::End.id()).unwrap()
-            }
-
-            writer.write_u32::<BigEndian>(value.len() as u32).unwrap();
-
-            for tag in value {
-                write_tag(writer, String::from(""), tag, false)?;
-            }
-        }
-        Tag::Compound(value) => {
-            for (name, tag) in value.tags {
-                write_tag(writer, name, tag, true)?;
-            }
-
-            writer.write_u8(Tag::End.id()).unwrap();
-        }
-        Tag::IntArray(value) => {
-            writer.write_u32::<BigEndian>(value.len() as u32).unwrap();
-
-            for v in value {
-                writer.write_i32::<BigEndian>(v).unwrap();
-            }
-        }
-        Tag::LongArray(value) => {
-            writer.write_u32::<BigEndian>(value.len() as u32).unwrap();
-
-            for v in value {
-                writer.write_i64::<BigEndian>(v).unwrap();
-            }
-        }
-        _ => return Err(TagError::UnknownType { type_id: tag.id() }),
-    }
-
-    Ok(())
-}
-
-fn write_string<W: Write>(writer: &mut W, value: String) {
-    writer.write_u16::<BigEndian>(value.len() as u16).unwrap();
-    writer.write(value.as_bytes()).unwrap();
-}
-
 #[test]
 fn test_compound_tag_i8() {
     let mut compound_tag = CompoundTag::new();
@@ -592,42 +363,4 @@ fn test_compound_tag_nested_compound_tag_vec() {
 
     assert_eq!(get_nested_compound_tag_1.get_str("str").unwrap(), "test");
     assert_eq!(get_nested_compound_tag_2.get_i32("i32").unwrap(), 222333111);
-}
-
-#[test]
-fn test_servers_read() {
-    let mut cursor = Cursor::new(include_bytes!("../test/servers.dat").to_vec());
-    let root_tag = read_compound_tag(&mut cursor).unwrap();
-
-    let servers = root_tag.get_compound_tag_vec("servers").unwrap();
-    assert_eq!(servers.len(), 1);
-
-    let server = servers[0];
-    let ip = server.get_str("ip").unwrap();
-    let name = server.get_str("name").unwrap();
-    let hide_address = server.get_bool("hideAddress").unwrap();
-
-    assert_eq!(ip, "localhost:25565");
-    assert_eq!(name, "Minecraft Server");
-    assert!(hide_address);
-}
-
-#[test]
-fn test_servers_write() {
-    let mut server = CompoundTag::new();
-
-    server.set_str("ip", "localhost:25565");
-    server.set_str("name", "Minecraft Server");
-    server.set_bool("hideAddress", true);
-
-    let mut servers = Vec::new();
-    servers.push(server);
-
-    let mut root_tag = CompoundTag::new();
-    root_tag.set_compound_tag_vec("servers", servers);
-
-    let mut vec = Vec::new();
-    write_compound_tag(&mut vec, root_tag).unwrap();
-
-    assert_eq!(vec, include_bytes!("../test/servers.dat").to_vec());
 }
